+++ src/main/java/blusunrize/immersiveengineering/common/register/IEBlocks.java
@@ -637,6 +637,7 @@ public final class IEBlocks
 		public static final BlockEntry<ClocheBlock> CLOCHE = new BlockEntry<>("cloche", METAL_PROPERTIES_NO_OCCLUSION, ClocheBlock::new);
 		public static final Map<IConveyorType<?>, BlockEntry<ConveyorBlock>> CONVEYORS = new HashMap<>();
 		public static final Map<EnumMetals, BlockEntry<ChuteBlock>> CHUTES = new EnumMap<>(EnumMetals.class);
+		public static final Map<DyeColor, BlockEntry<ChuteBlock>> DYED_CHUTES = new EnumMap<>(DyeColor.class);
 		public static final BlockEntry<AnyFacingEntityBlock<ElectromagnetBlockEntity>> ELECTROMAGNET = new BlockEntry<>(
 				"electromagnet", DEFAULT_METAL_PROPERTIES, p -> new AnyFacingEntityBlock<>(IEBlockEntities.ELECTROMAGNET, p)
 		);
@@ -648,7 +649,8 @@ public final class IEBlocks
 		{
 			for(EnumMetals metal : new EnumMetals[]{EnumMetals.IRON, EnumMetals.STEEL, EnumMetals.ALUMINUM, EnumMetals.COPPER})
 				CHUTES.put(metal, new BlockEntry<>("chute_"+metal.tagName(), METAL_PROPERTIES_DYNAMIC, ChuteBlock::new));
-
+			for(DyeColor dye : DyeColor.values())
+				DYED_CHUTES.put(dye, new BlockEntry<>("chute_colored_"+dye.getName(), METAL_PROPERTIES_DYNAMIC, ChuteBlock::new));
 		}
 
 		public static void initConveyors()